css: Merge all panels CSS files into a single/global stylesheet
A recurrent issue with the isolation of each panel stylesheet is
when a panels will unintentionally override the style of another
panel because a generic css selector is used.

Having all styles defined in the same stylesheet helps find these
accidental conflicts.

This also reduces the boilerplate code to load a style provider
for each panel.