display: Move Apply/Cancel buttons to the bottom of the pages
This also moves the error message to a toast, since we were relying on
the titlebar title to indicate that changes couldn't be applied.

This commit leaves the display settings on an intermediate
state towards the mockups at
https://gitlab.gnome.org/Teams/Design/settings-mockups/-/blob/4a5af0b6b7164a347a0e92833a1aa7232b62c1b2/displays/displays-aday.png

The next step should be introducing the new display settings switcher
that will replace our per-display subpages. This will allow for
removing the duplicated widgets and their bindings for the
display_settings_page and the toplevel page.

Fixes #3357