mouse: Adjust monochrome conversion script
Newer Blender outputs apparently contain either a different colorspace
or have more complete metadata. Converting these to monochrome results
in an output that is too light.

Correct for this by adding another conversion script for newer Blender
output that takes the video filter to be in sRGB. This now looks the
same as the old conversions again.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-control-center/-/merge_requests/3109>