system: Refactor and improve password generation
We don't have to load the entire file for every word in the password.
Instead, we load the whole file line by line into an array. This also
gives us the linecount automatically.

With that, we can also fix the slight bias in the random integer
generation, as just taking the modulo with the linecount is not
completely accurate. Using the glib function to generate a random
integer in a range does give uniform randomness.

(cherry picked from commit 50fb63a13f7bb10bf599843f9969b042caa8c4b6)