CcSplitRow: make it activatable so we can skip individual keyboard focus
As part of #3441 , let's make CcSplitRow respond to 'activate' signal
i.e. pressing space bar or ENTER to activate the individual items
inside it, this way we can set those items as "can-focus: FALSE"
so the keyboard focus chain skips them when pressing TAB (as they
can now be activated from the CcSplitRow itself).

Part of #3441