CcAvatarChooser: handle cc_crop_area_create_pixbuf() NULL return
cc_crop_area_create_pixbuf() can return NULL when
it fails to create GskRenderer, we were not handling
such error so we end up setting that failed image
as the user's image.

Catch the error and show a message dialog so the
user knows something went bad when cropping the image.

Fixes #3402 #3405 #3430