language-chooser: Port to CcLocaleRow
The CcLocaleRow port adds some accessibility improvements where
navigating with a keyboard will automatically select the row, as
the GtkListBox API intended.

The "double clicking" behavior that selects the language and closes
the dialog has been removed because it interfered with users selecting
the row with a keyboard. It also seemed to be an Easter egg feature
so it should be safe to remove.