background: Set a name for recent backgrounds
This ends up being the accessible name of the background item widget,
as spoken by the screen reader.

If there is no name, the widget will inherit the name of the first overlayed
widget with a name, which in this case is the "Remove Button". Thus, this
fixes the issue where highlighting the recent background gets announced as
"Remove Button" even if that button is not currently selected.