background: Do not use the list-item role for single preference rows
The "Style" and "Accent Color" entries are technically list-items of the
AdwPreferencesGroup, but we're not really using them as lists.
Because they're single children, the screen reader announces them as
"Style list with 1 item", which may be confusing as they contain multiple
selectable styles or accent colors.

These widgets instead behave like a group of toggle buttons, so mark them
with GTK_ACCESSIBLE_ROLE_GROUP. The screen reader now only announces the label.