background: Set accessible role for background widgets
We need to set an accessible role that has the checked/unchecked semantics,
so that the role is announced and that the checked state is announced.

Follow the other widgets in the background page (styles, accent colors) and
use the "toggle-button" role.

(cherry picked from commit 5515f00d1c7e78fbea60542c7496feb44fcc9ca7)