color: Use AdwDialog.force_close for dialog
We always create the dialog, even when it isn't shown. So we need to
force-close it to avoid calling adw_dialog_close on a dialog that wasn't
presented.

Also, the pointer nulling wasn't doing anything, as the dialog is
cleaned up automatically by the templating already.

Same as commit 2a197100

Related #3339