sound: Port CcDeviceComboRow to AdwComboRow
This removes deprecated code, and improves accessibility.

Because of the icon, this requires a custom factory implementation for
the list items. We reuse some of the code from AdwComboRow (and
CcBreakScheduleRow) to handle the selection icons properly.

We let the label ellipsize, but don't have a `max-width-chars`, so that
in the dropdown, the label can show fully. Otherwise long names can
never be fully viewed.

For the icon, we need to use fallbacks, as that is what the old
GtkCellRendererPixbuf also uses. Otherwise some icons (like for a
Bluetooth speaker) don't show up.

Closes #2855