style: Always explicitly pass template in blp files
It is default behavior of GTK to pass in the template object if no
object is specified for a signal. So, if the template object should be
the first argument of the callback, we can use `swapped`. But another
default behavior is that if an object is explicitly specified for a
signal, then it gets automatically swapped.

For clarity, it's therefore better to explicitly pass the template
object, as in Blueprint it is very clear that that is the main argument
of the signal callback then. Blueprint even warns if `swapped` is used
together with an explicit template object getting passed.

See
https://docs.gtk.org/gtk4/class.Builder.html#signal-handlers-and-function-pointers
https://gnome.pages.gitlab.gnome.org/blueprint-compiler/reference/objects.html#signal-handlers