style: Don't combine template and not-swapped in blp files
Similar to the previous commit, this is even more confusing. The passing
of `template` means auto-swap, but apparently there were a few cases
where that was undesirable, so `not-swapped` was added.

It's clearer to just not pass anything then, and rely on the "template
object is always passed as user_data" argument.