sound: Fix cut off of speaker test button border on hover
With the "circular" css style the speaker test buttons on the test wheel
get slightly cut off when hovered.

In
https://gitlab.gnome.org/GNOME/gnome-control-center/-/merge_requests/3046#note_2343521
Matthijs mentioned that Alice suggested we pack the buttons inside a bin
with border-radius 9999px and padding 2px to achieve the same effect
without the undesired cut off.

Also, .opaque is not desirable here as it offsets the focus ring like
for suggested action buttons. To keep the size similar, we increase the
padding on the button.

Obsoletes https://gitlab.gnome.org/GNOME/gnome-control-center/-/merge_requests/3046

Fixes #3336

Co-authored-by: Matthijs Velsink <mvelsink@gnome.org>