a11y: Fix issues in the manual network data
When an invalid IP address, network mask or gateway address is
typed, the only error notification is a change of the background
color to red, which is non-compliant with accessibity
requirements (don't rely on colors to specify a state).

This patch adds a `dialog-error-symbolic` icon in the wrong
field, and also uses `gtk_accessible_update_state()` to inform
when one of those fields with an invalid value.
It also moves this into ui-helpers, to make it available to
any GtkWidget field that uses it.

(cherry picked from commit a021824dbdc63c718931d41e8f8871f3d1aefdc7)