a11y: announce when an IP address is wrong
If a manual IP or netmask has invalid values, the only
notification to the user is setting the background in red color,
but no notification is sent through a screen reader.

This patch fixes this, by announcing if an IP or netmask field
is wrong as soon as the user moves out of, or into, the wrong
field.

(cherry picked from commit fd691b7b33730acfe738199204cee4bd0f30d04e)