wacom: Drop noinst calibrator helper
This small utility helped produce Xorg.conf snippets in order
to preserve tablet calibration widely to all Xorg environments
(e.g. outside of GNOME).

This utility was installed somewhere in libexec in the past, went
noinst when we went bolder in our pursue for Wayland, and now is
falling behind the times since we are deprecating X11 sessions
altogether.

This standalone utility has received a fair amount of use throughout
the years. If anyone misses it, it will be left as an exercise to
the reader to rescue it into a standalone repository.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-control-center/-/merge_requests/3182>