a11y: Specify what is being selected in avatar
When the user selects to change the user avatar, and the stock
avatars are shown, the screen reader just reads the name of the
first avatar ("Bycicle"), but gives no indication of what is
describing.

This patch adds a group name ("Stock avatar") to the list of
avatars available by default, thus giving context to the screen
reader.

It also adds a role to the avatar buttons in a separate
commit, to simplify backporting separately (the group name
requires a translation, thus is not suitable for backporting).

(cherry picked from commit fddf8c26588d5a9d59a682cd1c0000e48275580f)

Co-authored by: Felipe Borges <felipeborges@gnome.org>

This backport does not contain the newly translatable string.