about: fix NVIDIA GPU name display if get_renderer_from_switcheroo failed
If System Details failed to use switcheroo to iterate GPUs, it will
call gnome-control-center-print-renderer with only basic variables,
and it returns empty name if NVIDIA GPU is used with NVIDIA library
but the envvar is unset.  This also checks whether DRI_PRIME is set
before considering the name invalid.

Fixes #3474