wellbeing: Handle accounts managed by malcontent
Handle child accounts managed by the Parental Controls, by displaying a
banner that lets the them know some settings cannot be changed, with a
link to the Parental Controls application. The screen time limits
controls also become insensitive, and the change introduces a
requirement for malcontent 0.14.0, when wanting to use this optional
malcontent functionality.

Fixes: https://gitlab.gnome.org/GNOME/gnome-control-center/-/issues/3514