sound: Refactor and fix device selection
The port to AdwComboRow forgot to block the selection signal when adding
a new device (autoselection happens by default), so that the first
device in a list was always automatically enabled.

To fix this is not trivial, because the selection signal isn't handled
by the row, but by the sound panel code. To avoid having to block
signals cross-code, we refactor the CcDeviceComboRow code to handle the
signals internally.

Fixes #3506