split-row: Use AdwBin instead of GtkCheckButton
In later commits, we want to add focus rings and hover/pressed
states to parent widgets, which will inherently make the
GtkCheckButton widgets purely decorative. To be more precise,
this will make them an inappropriate candidate for displaying
radio indicators.

This commit ports to AdwBin and sets the appropriate roles and
states to the parent widgets, as a preparation for later commits.