ua-seeing-page: Hide screen reader configure row if unsupported
Orca started supporting a DBus interface at the same time as creating
the ShowPreferences method. So the simple way to check if the method is
support is to just call another implemented method (GetVersion) and see
if that call fails or not.

Also hide the Configure row if the proxy can't be created for some
reason.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-control-center/-/merge_requests/3215>