system, users, a11y: Announce password verify entry state changes
We check the validity of the passwords when both the password and verify
password entry widgets change. So instead of announcing both states to
the screen reader, let's only announce the one the state of the widget
that changed.

(cherry picked from commit ae224b5e70316f9e8579e5a55c86ffd162131edf)