keyboard: Update for new default activities overview shortcut
Mutter in 48 gained support for using both Super keys as overlay keys
and the default value of the overlay-key setting got changed to "Super"
accordingly. Because the keyboard shortcuts dialog was comparing the
overlay-key value to the old value of "Super_L" it was wrongly
considering the overview shortcut to be disabled.

Fix this by comparing to "Super" now, but also check for "Super_L" in
case somebody has toggled the setting before and as a result does not
have the new default value.

Related: https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4050
Closes: https://gitlab.gnome.org/GNOME/gnome-control-center/-/issues/3586