a11y: Replace "enable animations" with "reduced motion"
The "enable animations" setting is not a good accessibility toggle: it's
meant to be used as a performance toggle (e.g. for applications running
inside a VM or on a remote desktop). This is why #3099 required adding a
whole disclaimer about it.

The reduced-motion accessibility setting is going to be proxied by the
GNOME desktop portal to applications and toolkits.

See: gsettings-desktop-schemas!114