apps: fix "Type" being char wrapped in certain locales
The .ui file sets a content-width of 360 for this
AdwDialog but that gets exhausted due to long
options in the combo box for media types, and we
cannot guess an optimal content-width because the
combo box options are fetched from the system (i.e.
they are not static strings defined in source code)
and also the locale's translation of these options
may produce an even longer string (the case in #3478).

So the best approach is to use Pango markup around the
string to force it to not wrap, in that respect we
are lucky the 'title' property of AdwPreferencesRow
allows Pango markup by default.

Suggested by Matthijs Velsink, our fix use a code
approach using g_markup_printf_escaped() instead of
directly adding the pango markup to the property string
which would result in ugly escaped tags mixed with
the real string to be translated.

Fixes #3478