keyboard: Allow default Alternate Character keys
The user might prefer just not to mess with the keyboard layout — some
layouts do have unusual keys as 3rd level switchers —, so offer the
"layout default" option.

To do so, the logic requires an inversion, as the other use of this page
has a switch that is active to set a custom key, whereas here the switch
has to be inactive to set a custom key.

Furthermore, the label for the "unset" option isn't just "Disabled" as
for the Compose Key, but needs to be customized too. In this case, to
"Default".

Co-authored-by: Matthijs Velsink <mvelsink@gnome.org>

Closes #918, #3021