wifi: Actually turn off wifi when only managing a single device
In 757baa06 ("network, wifi: Don't disable wireless when toggling Wifi OFF")
the wifi toggle was changed to not actually turn off wifi, but to only
disconnect the currently selected wifi adapter from any network.
This was done to support the multi-wifi-adapter use-case.

This is poorly communicated in UI, and not what the user expects; it also
disables auto-connect when turning the wifi back on.
Let's restore the the good UX for the single-wifi-adapter case.

Related: https://gitlab.gnome.org/GNOME/gnome-control-center/-/issues/3227
Closes: https://gitlab.gnome.org/GNOME/gnome-control-center/-/issues/3460