keyboard: Add layout default to alternate char page
It adds value to state the default alternate characters key for a user.
However, best not to state it incorrectly. There are some (admittedly
obscure) layouts that have Level 3 on keys like Backslash or Capslock.

To test for this, we use GTK to retrieve the currently active keycodes
for the Level 3 shift, and then if anything other than Right Alt or the
custom key is returned, we do not show the layout default.

Otherwise, we stick to "Right Alt".

This needs to include the kernel keycodes, but those are available on
any system.