wellbeing: Add SPDX comment to bar chart documentation diagram
This was probably de-facto contributed as GPL-2.0-or-later as per the
top-level `COPYING` file. However, I think CC-BY-SA would be more
appropriate for a documentation image. I’m the only contributor, so
relicensing requires no permissions from any other contributors, if the
maintainers are happy to accept this.

I neglected to add the SPDX comment when first adding the file, which is
my mistake.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>