wacom: Show link to mouse settings for pucks
Traditionally, pucks were considered the same as mice in terms of
configuration. The not-so-recent updates to the Wacom panel to use
Adwaita inadvertently got rid of all distinctions for these devices,
making these show as styli with configurable buttons and pressure.

This was never intended, so switch back to showing a "link" to the
mouse configuration if a puck is found, now as an AdwActionRow in
the section of the related tablet.