keyboard: Set system keyboard mapping on single user systems
When there is a single user and that user has permission to set the
system keyboard mapping, it should do so when changing the user one.

This does not load the mapping from localed. And it does not provide a
way to set it on multi-user systems.

The code was mostly there. But some parts were not duplicated when
it was extracted from the region panel in
8996ffac88ae104781ccd3e5e842027ffeccc8a5.

Fixes #1968