wellbeing: Fix SPDX license identifier mismatch
Looks like I consistently typoed the SPDX license identifier in these
files when first writing them. The SPDX identifier is meant to match the
verbose license terms above, so fixing this mistake does not mean we’re
changing the license of these files (so should be straightforward to
merge). It just means we’re fixing the machine readable representation
of it.

Also change a typo in the verbose license terms to fix the LGPL version.
Realistically nobody uses LGPL 2.0 (as opposed to version 2.1), and it
certainly wasn’t my intention to use LGPL 2.0 when writing these files.
It probably arose from copying the license terms from a GPL 2.0 file and
not changing them properly.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Fixes: #3627