common: Add SPDX license header for cc-timelike-editor.blp
Populate the copyright lines from the git log for the file (going back
to `cc-time-editor.ui`), and populate the license from the top-level
`COPYING` file, which is what applies by default to all contributions.

This just clarifies the licensing of this `.blp` file, in line with the
other wellbeing `.blp` files in `panels/common/`.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>