power: fix infinite loop when switching battery charge modes
When toggling from "Preserve Battery Health" to "Maximize Charge" on
systems with multiple batteries, the UI hangs. The problem occurred on
my Lenovo T470, which has two batteries (one internal one, one
swappable).

I found the issue when inspecting the messages on DBus and the state was
constantly toggling, once I hit the "Maximize Charge" radio button.

  $ dbus-monitor --system "interface='org.freedesktop.UPower.Device'"
  [...]
  method call path=/org/freedesktop/UPower/devices/battery_BAT1; member=EnableChargeThreshold boolean true
  method call path=/org/freedesktop/UPower/devices/battery_BAT0; member=EnableChargeThreshold boolean false
  method call path=/org/freedesktop/UPower/devices/battery_BAT1; member=EnableChargeThreshold boolean false
  method call path=/org/freedesktop/UPower/devices/battery_BAT0; member=EnableChargeThreshold boolean true
  [repeating indefinitely...]

The cause of this toggling seems to be the following loop:

1. Initial state: Both batteries are set to Preserve Battery Health
   ("true")
2. When the UI switches to "Maximize Charge"
   EnableChargeThreshold(false) is sent to both batteries
3. BAT0 updates to false quickly
4. BAT1 is still true
5. up_client_changed() runs, sees BAT1 is still in "Maximize Charge"
   mode ("true") and toggles the UI back to "Preserve Battery Health".
   This triggers the callback again.
6. The callback sends EnableChargeThreshold(true) to both batteries
7. BAT0 gets set back to true, BAT1 finally updates to false
8. Now they're still out of sync and this creates the loop

I am not familiar with the code, but one option seems to be to block the
signal handler before calling gtk_check_button_set_active(), then
unblocking it after toggling all the batteries.